+++ app/src/main/java/com/example/healthcheck/navigation/destinations/ParamComposable.kt
@@ -11,7 +11,7 @@ import com.google.accompanist.navigation.animation.composable
 @ExperimentalMaterialApi
 fun NavGraphBuilder.paramComposable(
     navigateToHomeScreen: () -> Unit,
-    navigateToResultScreen: (Float) -> Unit
+    navigateToResultScreen: (Long) -> Unit
 
     ) {
     composable(
